requirements.txt
httpx>=0.27.0
pydantic>=2.5.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-dotenv>=1.0.0
# Optional — provides better embeddings than TF-IDF fallback
sentence-transformers>=2.7.0
groq